+++ src/ts/modules/dispatchEvent.ts
@@ -1,0 +1,4 @@
+export default <T>(name: string, params?: T) => {
+    const event = new CustomEvent(name, { detail: params })
+    window.dispatchEvent(event)
+}