+++ src/ts/modules/listenEvent.ts
@@ -1,0 +1,3 @@
+export default <T>(name: string, callback: (data: T) => void) => {
+    window.addEventListener(name, (e: any) => callback(e.detail))
+}